Remove duplicate entries of bin files in ignition
Scripts in /usr/local/bin were added to the bootstrap ignition twice,
resulting in the ones copied over to the nodes with 644 permission,
instead of expected 755. This commit ensures single entry is present in
the ignition file, with 755 mode

Since the scripts did not have executable permissions, the units
supposed to run the scripts needed to have explicit target for bash.
This is no longer needed, hence removed