Specify explicit cert for ingresscontroller
In 4.5, the console relies on the default ingress cert CA to access the
oauth endpoint. In hypershift, the oauth endpoint is served by a route
outside the cluster. For the console to use a valid CA, the cert used by
the default ingress must be able to validate both routes and the oauth
endpoint.