Use the parent cluster's domain as the base for DNS
For a parent cluster with domain mycluster.openshift.com, we were
assigning a child named child1 the ingress domain
apps.child1.openshift.com. This is causing issues updating the internal
DNS hosted zone because it starts with mycluster.openshift.com as the
base. With this change, we will now use
apps.child1.mycluster.openshift.com instead. This fixes the issue with
the private DNS zone.